+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: kpourcel <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 11:35:38 by kpourcel          #+#    #+#             */
-/*   Updated: 2023/11/03 17:12:22 by kpourcel         ###   ########.fr       */
+/*   Updated: 2023/11/05 12:37:21 by kpourcel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,14 +17,15 @@
 # include <stdarg.h>
 # include <stdio.h>
 # include <unistd.h>
+# include <limits.h>
 /*					Fonction utile 							*/
 int		ft_printf(const char *s, ...);
 int		ft_putchar(char c);
 int		ft_putstr(char *src);
 int		ft_putnbr2(int nb);
 int		ft_printunsigned_decimal(unsigned int nb);
-int		ft_printhexa_low(unsigned long nb);
-int		ft_printhexa_upper(unsigned long nb);
+int		ft_printhexa_low(unsigned int nb);
+int		ft_printhexa_upper(unsigned int nb);
 int		ft_printadresse(void *adresse);
 int		ft_printmodulo(char c);
